replaced: Newton Method - Binary search
Newton/Secand Method converges faster than binary search.

Additional, I changed the equations for the bullet tracing. In my opinion, it is not necessary to calculate the velocity magnitude. The friction depents on the velocity of each components of the velocity vector, not on a mixture of the magnitude and the component value.