Added the ability to defuse explosives (not entirely sure the current implementation is suitable, would be better to have a more involved defusal process.
Added the ability to set the timer of explosives, current implementation is just suitable but would be better to have the ability to select which type of trigger you would like for the explosive (specifically the ones that allow multiple trigger types). Then handling explosion via script to allow for a better timer system, exactly 15 seconds, exactly 3 minutes and 10 seconds, etc.