Update AGM_Explosives part I
So. The explosives will turn with player by default, enabling fast deploying a mine maze of 15 claymores.
If you hold Shift key and turn, the explosives won't turn with player, but keep their own directions.
Damn hard to describe in words, get part I and part II and try it out in game!